requirements.txt
fastapi 
fastapi[standard]
uvicorn
psycopg2-binary 
sqlalchemy 
pydantic 
python-dotenv
phidata
groq